+++ pincer/objects/user.py
@@ -25,9 +25,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from enum import IntEnum
-from typing import List, Optional
+from typing import Optional
 
-from .integration import Integration
 from ..utils import APIObject, APINullable, MISSING, Snowflake, convert
 
 
@@ -39,6 +38,7 @@ class PremiumTypes(IntEnum):
     NITRO_CLASSIC = 1
     NITRO = 2
 
+
 class VisibilityType(IntEnum):
     """
     The type of a connection visibility.
@@ -46,46 +46,6 @@ class VisibilityType(IntEnum):
     NONE = 0
     EVERYONE = 1
 
-@dataclass
-class Connection(APIObject):
-    """
-    The connection object that the user has attached.
-
-    :param id:
-        id of the connection account
-
-    :param name:
-        the username of the connection account
-
-    :param type:
-        the service of the connection (twitch, youtube)
-
-    :param revoked:
-        whether the connection is revoked
-
-    :param integrations:
-        an array of partial server integrations
-
-    :param verified:
-        whether the connection is verified
-
-    :param friend_sync:
-        whether friend sync is enabled for this connection
-
-    :param show_activity:
-        whether activities related to this connection
-        will be shown in presence updates
-    """
-    id: str
-    name: str
-    type: str
-    verified: bool
-    friend_sync: bool
-    show_activity: bool
-    visibility: VisibilityType
-
-    revoked: APINullable[bool] = MISSING
-    integrations: APINullable[List[Integration]] = MISSING
 
 @dataclass
 class User(APIObject):
